Tier 1+2: SDF text, UI layer, tween manager, GPU particles
Tier 1A - SDF Text Rendering:
- sdf_generator.rs: 8SSEDT dead-reckoning SDF gen, MSDF, shelf packing, disk cache
- sdf_atlas.rs: SdfAtlas with metrics, smoothing, outline/shadow/glow/bold
- sdf_batch.rs: SdfGlyphBatcher with per-instance effects
- sdf_glyph.vert/frag: SDF shaders with outline, shadow, glow, distortion

Tier 1B - Screen-Space UI Layer:
- ui_layer.rs: UiLayer with ortho projection, draw commands, 5 border styles
- ui_layer_renderer.rs: Converts commands to GlyphInstances, embedded shaders
- ui_primitives.rs: HP/MP/XP bars, panels, menus, tooltips, combat log, FPS overlay

Tier 2A - Tween Engine Wiring:
- tween_manager.rs: Central TweenManager with targets, delays, callbacks, tags
- game_tweens.rs: 25+ game presets (damage arcs, bar lerps, screen fades, etc.)

Tier 2B - GPU Particle Compute:
- gpu_particles.rs: GpuParticleSystem with double-buffered SSBOs, 10 engine types
- particle_update.comp: Compute shader with Lorenz/Rossler/Julia/etc + force fields
- particle_render.rs: Indirect draw from SSBO, LOD, depth layers

+++ src/deferred/gbuffer.rs
@@ -1764,7 +1764,7 @@ mod tests {
             [0.577, 0.577, 0.577],
         ];
         for n in &normals {
-            let len = (n[0] * n[0] + n[1] * n[1] + n[2] * n[2]).sqrt();
+            let len = (n[0] * n[0] + n[1] * n[1] + n[2] * n[2] as f32).sqrt();
             let normalized = [n[0] / len, n[1] / len, n[2] / len];
             let encoded = octahedral_encode(normalized);
             let decoded = octahedral_decode(encoded);
